Move AppRoutes back to routes
Fixes circular import issue between routes, App, topNavigation
