+++ src/application/applicationPreviewPage.tsx
@@ -9,7 +9,6 @@ import { Helmet } from 'react-helmet';
 
 import { RootState } from '../root/rootReducer';
 import {
-  AppRoutes,
   getClientErrorMessage,
   prepareApplicationForSubmission,
 } from './helpers';
@@ -23,7 +22,7 @@ import {
   FieldTypeMapping,
   UploadedFileMeta,
 } from './types';
-import { getRouteById } from '../root/routes';
+import { getRouteById, AppRoutes } from '../root/helpers';
 import { getPlotSearchFromFavourites } from '../favourites/helpers';
 import { Favourite } from '../favourites/types';
 import { PlotSearch } from '../plotSearch/types';
