+++ src/favouritesPage/components/favouriteCardMap.tsx
@@ -9,11 +9,10 @@ import { useNavigate } from 'react-router-dom';
 import { PlotSearchTarget } from '../../plotSearch/types';
 import { attachMapResizeObserver, getTargetCentroid } from '../../map/utils';
 import { initializeHelsinkiMap } from '../../map/utils';
-import { getRouteById } from '../../root/routes';
+import { getRouteById, AppRoutes } from '../../root/helpers';
 import { StandardMapLayer } from '../../map/StandardMapLayersControl';
 import { MapLayer, MapLayers } from '../../map/types';
 import MapReadyHandler from '../../map/MapReadyHandler';
-import { AppRoutes } from '../../application/helpers';
 
 interface Props {
   target: PlotSearchTarget;
