+++ src/mapSearch/mapPlotSearchOverlay.tsx
@@ -8,8 +8,7 @@ import { renderToStaticMarkup } from 'react-dom/server';
 import MapSymbol from './mapSymbol';
 import { FavouriteTarget } from '../favourites/types';
 import { useNavigate } from 'react-router';
-import { getRouteById } from '../root/routes';
-import { AppRoutes } from '../application/helpers';
+import { getRouteById, AppRoutes } from '../root/helpers';
 
 interface Props {
   plotSearchTargets: PlotSearchTarget[];
