+++ src/root/routes.tsx
@@ -15,56 +15,7 @@ import ApplicationPreviewPage from '../application/applicationPreviewPage';
 import AreaSearchLandingPage from '../areaSearch/areaSearchLandingPage';
 import AreaSearchApplicationRootPage from '../areaSearch/areaSearchApplicationRootPage';
 import DirectReservationPage from '../directReservation/directReservationPage';
-import { AppRoutes } from '../application/helpers';
-
-/**
- * Get route by id
- * @param {string} id
- * @returns {string}
- */
-export const getRouteById = (id: string): string => {
-  const routes = {
-    [AppRoutes.HOME]: '/',
-    [AppRoutes.PLOT_SEARCH_AND_COMPETITIONS]: '/tonttihaut-ja-kilpailut',
-    [AppRoutes.PLOT_SEARCH_AND_COMPETITIONS_TARGET]:
-      '/tonttihaut-ja-kilpailut/kohteet/',
-    [AppRoutes.OTHER_COMPETITIONS_AND_SEARCHES]: '/muut-kilpailut-ja-haut',
-    [AppRoutes.AREA_SEARCH_LANDING]: '/aluehaku',
-    [AppRoutes.AREA_SEARCH_APPLICATION_ROOT]: '/aluehaku/hakemus',
-    [AppRoutes.AREA_SEARCH_APPLICATION_AREA_SPEC]:
-      '/aluehaku/hakemus/alueen-maaritys',
-    [AppRoutes.AREA_SEARCH_APPLICATION_FORM]:
-      '/aluehaku/hakemus/tietojen-taytto',
-    [AppRoutes.AREA_SEARCH_APPLICATION_SUBMIT]: '/aluehaku/hakemus/lahetys',
-    [AppRoutes.AREA_SEARCH_APPLICATION_FORM_PREVIEW]:
-      '/aluehaku/hakemus/tietojen-tarkistus',
-    [AppRoutes.LEASES]: '/vuokraukset',
-    [AppRoutes.APPLICATIONS]: '/hakemukset',
-    [AppRoutes.MESSAGES]: '/viestit',
-    [AppRoutes.FAVOURITES]: '/suosikit',
-    [AppRoutes.OIDC_CALLBACK]: '/oidc/callback',
-    [AppRoutes.APPLICATION_ROOT]: '/hakemus',
-    [AppRoutes.APPLICATION_FORM]: '/hakemus/tietojen-taytto',
-    [AppRoutes.APPLICATION_PREVIEW]: '/hakemus/tietojen-tarkistus',
-    [AppRoutes.APPLICATION_SUBMIT]: '/hakemus/lahetys',
-    [AppRoutes.DIRECT_RESERVATION]: '/suoravaraus/',
-  };
-
-  return routes[id] ? routes[id] : '';
-};
-
-export const getPartialRouteById = (id: string, parentId: string): string => {
-  const target = getRouteById(id);
-  const parent = getRouteById(parentId);
-
-  if (!target.startsWith(parent)) {
-    throw new Error(
-      `Invalid route nesting pattern! ${target} is not a subroute of ${parent}.`,
-    );
-  }
-
-  return target.slice(parent.length);
-};
+import { getRouteById, getPartialRouteById, AppRoutes } from './helpers';
 
 const SiteRoutes = (): JSX.Element => {
   const RouteWithLoader = ({
